Hacked in a simple multiplication speed limited hash function

+++ src/catena-blake2b.c
@@ -3,7 +3,7 @@
 #include "hash.h"
 
 
-inline void __Hash1(const uint8_t *input, const uint8_t inputlen,
+inline void __Hash1(const uint8_t *input, const uint32_t inputlen,
 		      uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
@@ -15,8 +15,8 @@ inline void __Hash1(const uint8_t *input, const uint8_t inputlen,
 
 /***************************************************/
 
-inline void __Hash2(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
+inline void __Hash2(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
 		    uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
@@ -30,9 +30,9 @@ inline void __Hash2(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash3(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
+inline void __Hash3(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
 		    uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
@@ -45,10 +45,10 @@ inline void __Hash3(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash4(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
-		     const uint8_t *i4, const uint8_t i4len,
+inline void __Hash4(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
+		     const uint8_t *i4, const uint32_t i4len,
 		    uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
@@ -63,11 +63,11 @@ inline void __Hash4(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash5(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
-		    const uint8_t *i4, const uint8_t i4len,
-		    const uint8_t *i5, const uint8_t i5len,
+inline void __Hash5(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
+		    const uint8_t *i4, const uint32_t i4len,
+		    const uint8_t *i5, const uint32_t i5len,
 		    uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
