+++ src/catena-sha512.c
@@ -5,7 +5,7 @@
 
 /***************************************************/
 
-inline void __Hash1(const uint8_t *input, const uint8_t inputlen,
+inline void __Hash1(const uint8_t *input, const uint32_t inputlen,
 		      uint8_t hash[H_LEN])
 {
   SHA512_CTX ctx;
@@ -16,8 +16,8 @@ inline void __Hash1(const uint8_t *input, const uint8_t inputlen,
 
 /***************************************************/
 
-inline void __Hash2(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
+inline void __Hash2(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
 		    uint8_t hash[H_LEN])
 {
   SHA512_CTX ctx;
@@ -31,9 +31,9 @@ inline void __Hash2(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash3(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
+inline void __Hash3(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
 		    uint8_t hash[H_LEN])
 {
 
@@ -47,10 +47,10 @@ inline void __Hash3(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash4(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
-		     const uint8_t *i4, const uint8_t i4len,
+inline void __Hash4(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
+		     const uint8_t *i4, const uint32_t i4len,
 		    uint8_t hash[H_LEN])
 {
   SHA512_CTX ctx;
@@ -65,11 +65,11 @@ inline void __Hash4(const uint8_t *i1, const uint8_t i1len,
 
 /***************************************************/
 
-inline void __Hash5(const uint8_t *i1, const uint8_t i1len,
-		    const uint8_t *i2, const uint8_t i2len,
-		    const uint8_t *i3, const uint8_t i3len,
-		    const uint8_t *i4, const uint8_t i4len,
-		    const uint8_t *i5, const uint8_t i5len,
+inline void __Hash5(const uint8_t *i1, const uint32_t i1len,
+		    const uint8_t *i2, const uint32_t i2len,
+		    const uint8_t *i3, const uint32_t i3len,
+		    const uint8_t *i4, const uint32_t i4len,
+		    const uint8_t *i5, const uint32_t i5len,
 		    uint8_t hash[H_LEN])
 {
   SHA512_CTX ctx;
